+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.2.0] - UNRELEASED
+
+### Added
+- New API for returning the requested resources of a connected connector (/resources/requested)
+- New API for returning the enum-name of a policy for a given policy-pattern (/policy-pattern with request-body patter)
+
+### Fixes
+- Fix: For non-existing proxy configurations, "http://test" is no longer used as the default proxy configuration.
+- Workaround: Use BrokerUpdateMessage and send complete catalog instead of ResourceUpdateMessage when updating a resource at a Broker
+
 ## [4.1.0] - 2021-03-15
 
 ### Fixes